# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Universo Agustino API
  version: 1.0.0
  description: API v1 - marketplace académico UNSA (comunidad @unsa.edu.pe). Compra en custodia sin comisión.
servers:
  - url: http://localhost:4000/api/v1
...
paths:
...
  /orders/{id}/accept:
    post:
      tags: [orders]
      summary: El vendedor acepta compra o alquiler bazar PENDING → ACCEPTED (409 si expiró o no es bazar)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK }, '409': { description: BAD_STATE, NOT_RENTAL o RESERVATION_EXPIRED } }
...
  /orders/{id}/confirm-payment:
    post:
      tags: [orders]
      summary: El vendedor confirma el abono → ESCROW (custodia)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK }, '403': { description: FORBIDDEN }, '409': { description: BAD_STATE } }
  /orders/{id}/cancel:
    post:
      tags: [orders]
      summary: Cancela comprador o vendedor (PENDING/ACCEPTED/PAID) con motivo
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK (cancelledReason BUYER_CANCELLED/SELLER_REJECTED) }, '409': { description: BAD_STATE } }
...
  /orders/{id}/refund:
    post:
      tags: [orders]
      summary: Reembolso por moderación (moderator/admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK }, '403': { description: FORBIDDEN } }
...
  /sedes/schedule:
    get:
      tags: [staff]
      summary: Horario global America/Lima y hasta 60 feriados desde hoy; feriados y días sin horario están cerrados
      responses: { '200': { description: OK }, '201': { description: Creado }, '400': { description: VALIDATION }, '401': { description: Sin sesión }, '403': { description: Solo Técnico para cambios }, '404': { description: No encontrado }, '409': { description: Conflicto de horario, turno, sede o duplicado } }
...
  /staff/schedule:
    get:
      tags: [staff]
      summary: Consultar horario global y todos los feriados (equipo)
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: OK }, '201': { description: Creado }, '400': { description: VALIDATION }, '401': { description: Sin sesión }, '403': { description: Solo Técnico para cambios }, '404': { description: No encontrado }, '409': { description: Conflicto de horario, turno, sede o duplicado } }
  /staff/schedule/hours/{weekday}:
    put:
      tags: [staff]
      summary: Guardar horario diario dentro de 0–1440 minutos; no invalida turnos existentes
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: weekday, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [opens, closes], properties: { opens: { type: integer, minimum: 0, maximum: 1440 }, closes: { type: integer, minimum: 0, maximum: 1440 }, special: { type: boolean } } }
      responses: { '200': { description: OK }, '201': { description: Creado }, '400': { description: VALIDATION }, '401': { description: Sin sesión }, '403': { description: Solo Técnico para cambios }, '404': { description: No encontrado }, '409': { description: Conflicto de horario, turno, sede o duplicado } }
    delete:
      tags: [staff]
      summary: Cerrar día sin turnos existentes (Técnico)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: weekday, in: path, required: true, schema: { type: integer } }
      responses: { '200': { description: OK }, '201': { description: Creado }, '400': { description: VALIDATION }, '401': { description: Sin sesión }, '403': { description: Solo Técnico para cambios }, '404': { description: No encontrado }, '409': { description: Conflicto de horario, turno, sede o duplicado } }
  /staff/schedule/holidays:
    post:
      tags: [staff]
      summary: Crear feriado global de cierre todo el día (Técnico)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [date, reason], properties: { date: { type: string, format: date }, reason: { type: string, minLength: 3, maxLength: 200 } } }
      responses: { '200': { description: OK }, '201': { description: Creado }, '400': { description: VALIDATION }, '401': { description: Sin sesión }, '403': { description: Solo Técnico para cambios }, '404': { description: No encontrado }, '409': { description: Conflicto de horario, turno, sede o duplicado } }
  /staff/schedule/holidays/{id}:
    delete:
      tags: [staff]
      summary: Quitar feriado (Técnico)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: OK }, '201': { description: Creado }, '400': { description: VALIDATION }, '401': { description: Sin sesión }, '403': { description: Solo Técnico para cambios }, '404': { description: No encontrado }, '409': { description: Conflicto de horario, turno, sede o duplicado } }
...
  /payments/checkout/{orderId}:
    post:
      tags: [payments]
      summary: Crea la preferencia de Mercado Pago del pedido del comprador y devuelve la URL de pago
      security: [{ bearerAuth: [] }]
      parameters: [{ name: orderId, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: "{ data: { url } }" }, '403': { description: FORBIDDEN }, '404': { description: PROVIDER_OFF }, '409': { description: BAD_STATE o RESERVATION_EXPIRED } }
  /payments/webhook:
    post:
      tags: [payments]
      summary: Notificación de Mercado Pago. Valida x-signature (HMAC-SHA256), relee el pago y exige approved + external_reference + monto exacto → ESCROW
      responses: { '200': { description: "{ data: { outcome } }" }, '401': { description: BAD_SIGNATURE }, '404': { description: PROVIDER_OFF } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...